+++ lib/lib.gni
@@ -293,6 +293,7 @@ libjxl_dec_sources = [
 libjxl_enc_sources = [
     "jxl/butteraugli/butteraugli.cc",
     "jxl/butteraugli/butteraugli.h",
+    "jxl/enc_Knuth_partition.h",
     "jxl/enc_ac_strategy.cc",
     "jxl/enc_ac_strategy.h",
     "jxl/enc_adaptive_quantization.cc",
@@ -352,6 +353,15 @@ libjxl_enc_sources = [
     "jxl/enc_icc_codec.h",
     "jxl/enc_image_bundle.cc",
     "jxl/enc_image_bundle.h",
+    "jxl/enc_jpeg_cluster.h",
+    "jxl/enc_jpeg_frame.cc",
+    "jxl/enc_jpeg_frame.h",
+    "jxl/enc_jpeg_histogram.h",
+    "jxl/enc_jpeg_opt_data.cc",
+    "jxl/enc_jpeg_opt_data.h",
+    "jxl/enc_jpeg_refine.h",
+    "jxl/enc_jpeg_search.h",
+    "jxl/enc_jpeg_threshold.h",
     "jxl/enc_linalg.cc",
     "jxl/enc_linalg.h",
     "jxl/enc_lz77.cc",
